+++ AddOns/REPorter/ElvUI.lua
@@ -1,17 +1,13 @@
-if not AddOnSkins then return end
-local AS = unpack(AddOnSkins)
-if not AS:CheckAddOn("REPorter") then return end
+if ElvUI then
+  local E = unpack(ElvUI)
+  local S = E:GetModule("Skins")
 
-function AS:REPorter()
-  AS:StripTextures(REPorterFrameBorder)
-  AS:SkinBackdropFrame(REPorterFrameBG, "Transparent")
-  REPorterFrameBG:SetOutside(nil, 2, 2)
-  REPorterFrameBorderResize:Point("BOTTOMRIGHT", REPorterFrameBorder, "BOTTOMRIGHT", 2, -2)
-
-  AS:SkinFrame(REPorterBar)
-  for _, i in pairs({"B1", "B2", "B3", "B4", "B5", "B6", "B7", "B8"}) do
-    AS:SkinButton(_G["REPorterBar"..i])
+  local function Skin_REPorter()
+    REPorterFrameBorder:StripTextures()
+    S:HandleFrame(REPorterFrameBG, true)
+    REPorterFrameBG:SetOutside(nil, 2, 2)
+    REPorterFrameBorderResize:Point("BOTTOMRIGHT", REPorterFrameBorder, "BOTTOMRIGHT", 2, -2)
   end
-end
 
-AS:RegisterSkin("REPorter", AS.REPorter)
+  S:AddCallbackForAddon("REPorter", "REPorter", Skin_REPorter)
+end