+++ AddOns/REPorter/REPorter.lua
@@ -1,6 +1,5 @@
 local _, RE = ...
 local L = LibStub("AceLocale-3.0"):GetLocale("REPorter")
-local LBS = LibStub("LibBabble-SubZone-3.0"):GetReverseLookupTable()
 local TIMER = LibStub("AceTimer-3.0")
 REPorter = RE
 
@@ -29,7 +28,6 @@ local GetNumBattlefieldFlagPositions = GetNumBattlefieldFlagPositions
 local GetBattlefieldFlagPosition = C_PvP.GetBattlefieldFlagPosition
 local GetNumBattlefieldVehicles = GetNumBattlefieldVehicles
 local GetBattlefieldVehicleInfo = C_PvP.GetBattlefieldVehicleInfo
-local GetSubZoneText = GetSubZoneText
 local GetClassColor = GetClassColor
 local GetRaidTargetIndex = GetRaidTargetIndex
 local GetDoubleStatusBarWidgetVisualizationInfo = C_UIWidgetManager.GetDoubleStatusBarWidgetVisualizationInfo
@@ -38,15 +36,10 @@ local UnitClass = UnitClass
 local UnitExists = UnitExists
 local UnitIsUnit = UnitIsUnit
 local UnitAffectingCombat = UnitAffectingCombat
-local UnitHealth = UnitHealth
-local UnitHealthMax = UnitHealthMax
 local UnitIsDeadOrGhost = UnitIsDeadOrGhost
 local UnitGroupRolesAssigned = UnitGroupRolesAssigned
-local SendChatMessage = SendChatMessage
 local SendAddonMessage = C_ChatInfo.SendAddonMessage
-local CombatLogGetCurrentEventInfo = CombatLogGetCurrentEventInfo
 local RegisterAddonMessagePrefix = C_ChatInfo.RegisterAddonMessagePrefix
-local Contains = tContains
 
 local AV = 91
 local WG = 1339
@@ -84,20 +77,13 @@ RE.POIList = {}
 RE.VignetteInfo = {}
 RE.VignettePosition = {}
 RE.PinTextures = {}
-RE.ClickedPOI = ""
 
 RE.DefaultTimer = 60
 RE.CareAboutNodes = false
 RE.CareAboutPoints = false
-RE.CareAboutGates = false
 RE.CareAboutFlags = false
 RE.CareAboutVehicles = false
 RE.PlayedFromStart = true
-RE.IoCAllianceGateName = ""
-RE.IoCHordeGateName = ""
-RE.IoCGateHealth = 2400000
-RE.IoCGateEstimator = {}
-RE.IoCGateEstimatorText = ""
 RE.SMEstimatorText = ""
 RE.SMEstimatorReport = ""
 RE.EstimatorTicks = {10000, 10000}
@@ -114,7 +100,7 @@ RE.BlinkPOIValue = 0.3
 RE.BlinkPOIUp = true
 
 RE.FoundNewVersion = false
-RE.AddonVersionCheck = 21011
+RE.AddonVersionCheck = 30000
 RE.ScreenHeight, RE.ScreenWidth = UIParent:GetCenter()
 
 RE.MapSettings = {
@@ -134,13 +120,6 @@ RE.MapSettings = {
 	[CI] = {["PlayerNumber"] = 10},
 	[ASH] = {["PlayerNumber"] = 40},
 	[DR] = {["PlayerNumber"] = 10, ["WidgetID"] = 1687}
-}
-RE.ZonesWithoutSubZones = {
-	[SM] = true,
-	[TOK] = true,
-	[TMVS] = true,
-	[CI] = true,
-	[DR] = true
 }
 RE.POICaptureStatus = {
 	[4] = FACTION_ALLIANCE, -- Graveyard
@@ -233,26 +212,14 @@ RE.BackdropB = {
 	tile = true,
 	tileSize = 32,
 }
-RE.BackdropC = {
-	edgeFile = "Interface\\FriendsFrame\\UI-Toast-Border",
-	bgFile = "Interface\\TutorialFrame\\TutorialFrameBackground",
-	tile = true,
-	tileSize = 32,
-	edgeSize = 12,
-	insets = { left = 5, right = 5, top = 5, bottom = 5 },
-}
 
 RE.DefaultConfig = {
 	profile = {
-		BarHandle = 11,
-		BarX = RE.ScreenHeight,
-		BarY = RE.ScreenWidth,
 		Locked = false,
 		Opacity = 0.75,
 		HideMinimap = false,
 		DisplayMarks = false,
 		DisplayHealers = false,
-		UseRaidWarnings = false,
 		Map = {
 			[AB] = {["wx"] = RE.ScreenHeight, ["wy"] = RE.ScreenWidth, ["ww"] = 325, ["wh"] = 325, ["mx"] = 16, ["my"] = -77, ["ms"] = 1},
 			[WG] = {["wx"] = RE.ScreenHeight, ["wy"] = RE.ScreenWidth, ["ww"] = 280, ["wh"] = 460, ["mx"] = -5, ["my"] = -38, ["ms"] = 1},
@@ -272,20 +239,6 @@ RE.DefaultConfig = {
 			[DR] = {["wx"] = RE.ScreenHeight, ["wy"] = RE.ScreenWidth, ["ww"] = 270, ["wh"] = 330, ["mx"] = 15, ["my"] = -40, ["ms"] = 1}
 		}
 	}
-}
-RE.ReportBarAnchor = {
-	[1] = {"BOTTOMLEFT", "BOTTOMRIGHT"},
-	[2] = {"LEFT", "RIGHT"},
-	[3] = {"TOPLEFT", "TOPRIGHT"},
-	[4] = {"BOTTOMRIGHT", "BOTTOMLEFT"},
-	[5] = {"RIGHT", "LEFT"},
-	[6] = {"TOPRIGHT", "TOPLEFT"},
-	[7] = {"BOTTOMLEFT", "TOPLEFT"},
-	[8] = {"BOTTOM", "TOP"},
-	[9] = {"BOTTOMRIGHT", "TOPRIGHT"},
-	[10] = {"TOPLEFT", "BOTTOMLEFT"},
-	[11] = {"TOP", "BOTTOM"},
-	[12] = {"TOPRIGHT", "BOTTOMRIGHT"}
 }
 RE.AceConfig = {
 	type = "group",
@@ -336,47 +289,12 @@ RE.AceConfig = {
 					set = function(_, val) RE.Settings.profile.DisplayHealers = val; RE:UpdateConfig() end,
 					get = function(_) return RE.Settings.profile.DisplayHealers end
 				},
-				UseRaidWarnings = {
-					name = L["Use raid warnings"],
-					desc = L["When checked alerts will also be sent as raid warnings."],
-					type = "toggle",
-					width = "full",
-					order = 6,
-					set = function(_, val) RE.Settings.profile.UseRaidWarnings = val; RE:UpdateConfig() end,
-					get = function(_) return RE.Settings.profile.UseRaidWarnings end
-				},
-				BarHandle = {
-					name = L["Report bar location"],
-					desc = L["Anchor point of a bar with quick report buttons."],
-					type = "select",
-					width = "double",
-					order = 7,
-					values = {
-						[1] = L["Right bottom"],
-						[2] = L["Right"],
-						[3] = L["Right top"],
-						[4] = L["Left bottom"],
-						[5] = L["Left"],
-						[6] = L["Left top"],
-						[7] = L["Top left"],
-						[8] = L["Top"],
-						[9] = L["Top right"],
-						[10] = L["Bottom left"],
-						[11] = L["Bottom"],
-						[12] = L["Bottom right"],
-						[13] = L["Standalone - Horizontal"],
-						[14] = L["Standalone - Vertical"],
-						[15] = L["Hidden"]
-					},
-					set = function(_, val) RE.Settings.profile.BarHandle = val; RE.Settings.profile.BarX, RE.Settings.profile.BarY = REPorterBar:GetCenter(); RE:UpdateConfig() end,
-					get = function(_) return RE.Settings.profile.BarHandle end
-				},
 				MapSettings = {
 					name = BATTLEGROUND,
 					desc = L["Map position is saved separately for each battleground."],
 					type = "select",
 					width = "double",
-					order = 8,
+					order = 6,
 					disabled = function(_) if select(2, IsInInstance()) == "pvp" then return true else return false end end,
 					values = {
 						[AB] = GetMapInfo(AB).name,
@@ -404,7 +322,7 @@ RE.AceConfig = {
 					desc = L["This option control map size."],
 					type = "range",
 					width = "double",
-					order = 9,
+					order = 7,
 					min = 0.5,
 					max = 1.5,
 					step = 0.05,
@@ -416,7 +334,7 @@ RE.AceConfig = {
 					desc = L["This option control map transparency."],
 					type = "range",
 					width = "double",
-					order = 10,
+					order = 8,
 					isPercent = true,
 					min = 0.1,
 					max = 1,
@@ -450,7 +368,6 @@ function RE:ShortTime(TimeRaw)
 	local TimeSec = floor(TimeRaw % 60)
 	local TimeMin = floor(TimeRaw / 60)
 	if TimeSec < 10 then
-        ---@diagnostic disable-next-line: cast-local-type
 		TimeSec = "0" .. TimeSec
 	end
 	return TimeMin .. ":" .. TimeSec
@@ -480,26 +397,6 @@ function RE:FramesOverlap(frameA, frameB)
 	and (frameB:GetBottom() * sB) < (frameA:GetTop() * sA)
 end
 
-function RE:POIDropDown(root)
-	local submenu = root:CreateButton(L["Incoming"])
-	submenu:CreateButton("1", function() RE:SmallButton(1, true) end)
-	submenu:CreateButton("2", function() RE:SmallButton(2, true) end)
-	submenu:CreateButton("3", function() RE:SmallButton(3, true) end)
-	submenu:CreateButton("4", function() RE:SmallButton(4, true) end)
-	submenu:CreateButton("5", function() RE:SmallButton(5, true) end)
-	submenu:CreateButton("5+", function() RE:SmallButton(6, true) end)
-    root:CreateButton(HELP_LABEL, function() RE:BigButton(true, true) end)
-	root:CreateButton(L["Clear"], function() RE:BigButton(false, true) end)
-	root:CreateDivider()
-	root:CreateButton(ATTACK, function() RE:ReportDropDownClick("Attack") end)
-	root:CreateButton(L["Guard"], function() RE:ReportDropDownClick("Guard") end)
-	root:CreateButton(L["Heavily defended"], function() RE:ReportDropDownClick("Heavily defended") end)
-	root:CreateButton(L["Losing"], function() RE:ReportDropDownClick("Losing") end)
-	root:CreateDivider()
-	root:CreateButton(L["On my way"], function() RE:ReportDropDownClick("On my way") end)
-	root:CreateButton(L["Report status"], function() RE:ReportDropDownClick("") end)
-end
-
 function RE:CreatePOI(index)
 	local frameMain = CreateFrame("Frame", "REPorterFrameCorePOI"..index, REPorterFrameCorePOI)
 	frameMain:SetFrameLevel(10)
@@ -507,7 +404,6 @@ function RE:CreatePOI(index)
 	frameMain:SetHeight(RE.POIIconSize)
 	frameMain:SetScript("OnEnter", function(self) RE:UnitOnEnterPOI(self) end)
 	frameMain:SetScript("OnLeave", function() GameTooltip:Hide() end)
-	frameMain:SetScript("OnMouseDown", function(self) RE:OnClickPOI(self) end)
 	local texture = frameMain:CreateTexture(frameMain:GetName().."Texture", "BORDER")
 	texture:SetPoint("CENTER", frameMain, "CENTER")
 	texture:SetWidth(RE.POIIconSize - 13)
@@ -574,23 +470,6 @@ function RE:OnLoad(self)
 	RE.FlagsPool = CreateFramePool("FRAME", REPorterFrameCore, "REPorterFlagTemplate")
 end
 
-function RE:OnEnterBar(_)
-	TIMER:CancelTimer(RE.TimerBar)
-	REPorterBar:SetAlpha(RE.Settings.profile.Opacity)
-end
-
-function RE:OnLeaveBar(_)
-	if not REPorterBar:IsMouseOver() then
-		TIMER:CancelTimer(RE.TimerBar)
-		RE.TimerBar = TIMER:ScheduleTimer(function() REPorterBar:SetAlpha(0.25) end, 0.5)
-	end
-end
-
-function RE:OnLeave(_)
-	TIMER:CancelTimer(RE.TimerDropDown)
-	RE.TimerDropDown = TIMER:ScheduleTimer(function() CloseDropDownMenus() end, 3)
-end
-
 function RE:OnDragStart(_)
 	REPorterFrameCore:ClearAllPoints()
 	REPorterFrameCore:SetPoint("CENTER", REPorterFrameCoreAnchor, "CENTER")
@@ -624,7 +503,7 @@ function RE:OnMouseWheel(delta)
 	newscale = RE:Round(newscale, 2)
 	REPorterFrameCore:SetScale(newscale)
 	if SettingsPanel:IsShown() then
-		RE.ConfigFrame.obj.children[1].children[1].children[9]:SetValue(newscale)
+		RE.ConfigFrame.obj.children[1].children[1].children[7]:SetValue(newscale)
 	end
 end
 
@@ -639,19 +518,10 @@ function RE:OnEvent(self, event, ...)
 
 		LibStub("AceConfigRegistry-3.0"):RegisterOptionsTable("REPorter", RE.AceConfig)
 		RE.ConfigFrame, RE.Settings.profile.categoryID = LibStub("AceConfigDialog-3.0"):AddToBlizOptions("REPorter", L["REPorter"]) -- 自行修改
-		SettingsPanel:HookScript("OnHide", function() RE:HideDummyMap(true) end)
+		hooksecurefunc(SettingsPanel, "Close", function(self) RE:HideDummyMap(true) end)
 		RE:UpdateConfig()
 
 		RegisterAddonMessagePrefix("REPorter")
-		BINDING_NAME_REPORTERINC1 = L["Incoming"].." 1"
-		BINDING_NAME_REPORTERINC2 = L["Incoming"].." 2"
-		BINDING_NAME_REPORTERINC3 = L["Incoming"].." 3"
-		BINDING_NAME_REPORTERINC4 = L["Incoming"].." 4"
-		BINDING_NAME_REPORTERINC5 = L["Incoming"].." 5"
-		BINDING_NAME_REPORTERINC6 = L["Incoming"].." 5+"
-		BINDING_NAME_REPORTERHELP = HELP_LABEL
-		BINDING_NAME_REPORTERCLEAR = L["Clear"]
-		REPorterBar:SetHitRectInsets(-5, -5, -5, -5)
 		REPorterFrameClip:SetClipsChildren(true)
 		REPorterFrameCoreUP.excludedMouseOverUnits = {}
 		REPorterFrameCoreUP:SetMouseOverUnitExcluded("player", true)
@@ -661,8 +531,6 @@ function RE:OnEvent(self, event, ...)
 		for i=1, RE.POINumber do
 			RE:CreatePOI(i)
 		end
-
-		RE.IsSkinned = AddOnSkins and AddOnSkins[1]:CheckOption("REPorter") or false
 
 		self:UnregisterEvent("ADDON_LOADED")
 	elseif event == "CHAT_MSG_ADDON" and ... == "REPorter" then
@@ -675,52 +543,6 @@ function RE:OnEvent(self, event, ...)
 				RE.FoundNewVersion = true
 			end
 		end
-	elseif event == "COMBAT_LOG_EVENT_UNFILTERED" and next(RE.POINodes) ~= nil then
-		local _, e, _, _, _, _, _, guid, _, _, _, _, _, _, damage = CombatLogGetCurrentEventInfo()
-		if e ~= "SPELL_BUILDING_DAMAGE" then return end
-
-        ---@diagnostic disable-next-line: need-check-nil
-		local gateID = guid:match("%-(%d-)%-%x-$")
-		if gateID == "195496" then -- Horde East
-			RE.POINodes[RE.IoCHordeGateName.." - "..L["East"]].health = RE.POINodes[RE.IoCHordeGateName.." - "..L["East"]].health - damage
-			if RE.POINodes[RE.IoCHordeGateName.." - "..L["East"]].health < RE.IoCGateEstimator[FACTION_HORDE] then
-				RE.IoCGateEstimator[FACTION_HORDE] = RE.POINodes[RE.IoCHordeGateName.." - "..L["East"]].health
-			end
-		elseif gateID == "195494" then -- Horde Central
-			RE.POINodes[RE.IoCHordeGateName.." - "..L["Front"]].health = RE.POINodes[RE.IoCHordeGateName.." - "..L["Front"]].health - damage
-			if RE.POINodes[RE.IoCHordeGateName.." - "..L["Front"]].health < RE.IoCGateEstimator[FACTION_HORDE] then
-				RE.IoCGateEstimator[FACTION_HORDE] = RE.POINodes[RE.IoCHordeGateName.." - "..L["Front"]].health
-			end
-		elseif gateID == "195495" then -- Horde West
-			RE.POINodes[RE.IoCHordeGateName.." - "..L["West"]].health = RE.POINodes[RE.IoCHordeGateName.." - "..L["West"]].health - damage
-			if RE.POINodes[RE.IoCHordeGateName.." - "..L["West"]].health < RE.IoCGateEstimator[FACTION_HORDE] then
-				RE.IoCGateEstimator[FACTION_HORDE] = RE.POINodes[RE.IoCHordeGateName.." - "..L["West"]].health
-			end
-		elseif gateID == "195700" then -- Alliance East
-			RE.POINodes[RE.IoCAllianceGateName.." - "..L["East"]].health = RE.POINodes[RE.IoCAllianceGateName.." - "..L["East"]].health - damage
-			if RE.POINodes[RE.IoCAllianceGateName.." - "..L["East"]].health < RE.IoCGateEstimator[FACTION_ALLIANCE] then
-				RE.IoCGateEstimator[FACTION_ALLIANCE] = RE.POINodes[RE.IoCAllianceGateName.." - "..L["East"]].health
-			end
-		elseif gateID == "195698" then -- Alliance Center
-			RE.POINodes[RE.IoCAllianceGateName.." - "..L["Front"]].health = RE.POINodes[RE.IoCAllianceGateName.." - "..L["Front"]].health - damage
-			if RE.POINodes[RE.IoCAllianceGateName.." - "..L["Front"]].health < RE.IoCGateEstimator[FACTION_ALLIANCE] then
-				RE.IoCGateEstimator[FACTION_ALLIANCE] = RE.POINodes[RE.IoCAllianceGateName.." - "..L["Front"]].health
-			end
-		elseif gateID == "195699" then -- Alliance West
-			RE.POINodes[RE.IoCAllianceGateName.." - "..L["West"]].health = RE.POINodes[RE.IoCAllianceGateName.." - "..L["West"]].health - damage
-			if RE.POINodes[RE.IoCAllianceGateName.." - "..L["West"]].health < RE.IoCGateEstimator[FACTION_ALLIANCE] then
-				RE.IoCGateEstimator[FACTION_ALLIANCE] = RE.POINodes[RE.IoCAllianceGateName.." - "..L["West"]].health
-			end
-		end
-
-		if RE.IoCGateEstimator[FACTION_HORDE] < RE.IoCGateEstimator[FACTION_ALLIANCE] then
-			RE.IoCGateEstimatorText = "|cFF00A9FF"..RE:Round((RE.IoCGateEstimator[FACTION_HORDE] / RE.IoCGateHealth) * 100, 0).."%|r"
-		elseif RE.IoCGateEstimator[FACTION_HORDE] > RE.IoCGateEstimator[FACTION_ALLIANCE] then
-			RE.IoCGateEstimatorText = "|cFFFF141D"..RE:Round((RE.IoCGateEstimator[FACTION_ALLIANCE] / RE.IoCGateHealth) * 100, 0).."%|r"
-		else
-			RE.IoCGateEstimatorText = ""
-		end
-		REPorterFrameEstimatorText:SetText(RE.IoCGateEstimatorText)
 	elseif event == "PLAYER_ENTERING_WORLD" or event == "ZONE_CHANGED_NEW_AREA" then
 		local instance = select(2, IsInInstance())
 		if RE.CurrentMap ~= -1 then
@@ -728,7 +550,6 @@ function RE:OnEvent(self, event, ...)
 		end
 		if instance ~= "pvp" then
 			REPorterFrame:Hide()
-			REPorterBar:Hide()
 			REPorterFrameEstimator:Hide()
 		end
 		if instance == "pvp" and RE.CurrentMap == -1 then
@@ -873,7 +694,6 @@ function RE:OnPOIUpdate()
 			if RE.CurrentMap == IOC then
 				RE.POIInfo.gate = false
 				if RE:CheckCoordinates(x, y, 421, -401) then
-					RE.IoCAllianceGateName = RE.POIInfo.name
 					RE.POIInfo.name = RE.POIInfo.name.." - "..L["East"]
 					RE.POIInfo.translatedName = "Alliance Gate - East"
 					RE.POIInfo.gate = true
@@ -889,7 +709,6 @@ function RE:OnPOIUpdate()
 					RE.POIInfo.gate = true
 					y = y + 15
 				elseif RE:CheckCoordinates(x, y, 380, -165) then
-					RE.IoCHordeGateName = RE.POIInfo.name
 					RE.POIInfo.name = RE.POIInfo.name.." - "..L["Front"]
 					RE.POIInfo.translatedName = "Horde Gate - Front"
 					RE.POIInfo.gate = true
@@ -946,15 +765,13 @@ function RE:OnPOIUpdate()
 					colorOverride = {0, 1, 0}
 				end
 			elseif RE.CurrentMap == BFW then
-				if Contains(RE.BFWWalls, RE.POIInfo.textureIndex) then
+				if tContains(RE.BFWWalls, RE.POIInfo.textureIndex) then
 					RE.POIInfo.name = RE.POIInfo.name.." "..RE.POIInfo.areaPoiID
 				end
 			end
 			if RE.POINodes[RE.POIInfo.name] == nil then
 				RE.POINodes[RE.POIInfo.name] = {["id"] = i, ["poiID"] = RE.POIInfo.areaPoiID, ["name"] = RE.POIInfo.name, ["status"] = RE.POIInfo.description, ["x"] = x, ["y"] = y, ["texture"] = RE.POIInfo.textureIndex, ["active"] = true}
 				if RE.CurrentMap == IOC and RE.POIInfo.gate then
-					RE.POINodes[RE.POIInfo.name].health = RE.IoCGateHealth
-					RE.POINodes[RE.POIInfo.name].maxHealth = RE.IoCGateHealth
 					RE.POINodes[RE.POIInfo.name].translatedName = RE.POIInfo.translatedName
 				elseif RE.CurrentMap == SS and RE.PlayedFromStart then
 					RE:NodeChange(RE.POIInfo.textureIndex, RE.POIInfo.name)
@@ -995,19 +812,6 @@ end
 function RE:OnUpdate(elapsed)
 	if RE.UpdateTimer < 0 then
 		RE:BlinkPOI()
-		if not RE.ZonesWithoutSubZones[RE.CurrentMap] then
-			local subZoneName = GetSubZoneText()
-			if subZoneName and subZoneName ~= "" then
-				for _, i in pairs({"B1", "B2", "B3", "B4", "B5", "B6", "B7", "B8"}) do
-					_G["REPorterBar"..i]:Enable()
-				end
-			else
-				for _, i in pairs({"B1", "B2", "B3", "B4", "B5", "B6", "B7", "B8"}) do
-					_G["REPorterBar"..i]:Disable()
-				end
-			end
-		end
-
 		if RE.NeedRefresh then
 			RE.NeedRefresh = false
 			REPorterFrameCoreUP:ClearUnits()
@@ -1021,11 +825,7 @@ function RE:OnUpdate(elapsed)
 					texture = "Interface\\Addons\\REPorter\\Textures\\BlipNormal"
 					local r, g, b = GetClassColor(select(2, UnitClass(unit)))
 					if UnitAffectingCombat(unit) then
-						if (UnitHealth(unit) / UnitHealthMax(unit)) * 100 < 26 then
-							texture = "Interface\\Addons\\REPorter\\Textures\\BlipDying"
-						else
-							texture = "Interface\\Addons\\REPorter\\Textures\\BlipCombat"
-						end
+						texture = "Interface\\Addons\\REPorter\\Textures\\BlipCombat"
 					elseif UnitIsDeadOrGhost(unit) then
 						texture = "Interface\\Addons\\REPorter\\Textures\\BlipDead"
 						r, g, b = r * 0.35, g * 0.35, b * 0.35
@@ -1128,7 +928,7 @@ function RE:OnUpdate(elapsed)
 				    else
 				      if RE.CurrentMap == SS then
 				        _G[battlefieldPOIName.."TextureBG"]:SetColorTexture(0,1,0,0.3)
-							elseif RE.CurrentMap == CI or (RE.CurrentMap == BFW and Contains(RE.BFWWalls, v.texture)) then
+							elseif RE.CurrentMap == CI or (RE.CurrentMap == BFW and tContains(RE.BFWWalls, v.texture)) then
 								_G[battlefieldPOIName.."TextureBG"]:SetColorTexture(0,0,0,0)
 				      else
 				        _G[battlefieldPOIName.."TextureBG"]:SetColorTexture(0,0,0,0.3)
@@ -1136,21 +936,9 @@ function RE:OnUpdate(elapsed)
 				    end
 				    _G[battlefieldPOIName.."TextureBG"]:SetWidth(RE.POIIconSize)
 				    _G[battlefieldPOIName.."TextureBGofBG"]:Hide()
-				    if RE.CareAboutGates and v.health and v.health > 0 then
-				      _G[battlefieldPOIName.."TextureBGTop1"]:Hide()
-				      _G[battlefieldPOIName.."TextureBGTop2"]:Show()
-				      _G[battlefieldPOIName.."TextureBGTop2"]:SetWidth((v.health/v.maxHealth) * RE.POIIconSize)
-				      if RE.PlayedFromStart then
-								_G[battlefieldPOIName.."TimerCaption"]:SetText(RE:Round((v.health/v.maxHealth)*100, 0).."%")
-				      else
-				        _G[battlefieldPOIName.."TimerCaption"]:SetText("|cFFFF141D"..RE:Round((v.health/v.maxHealth)*100, 0).."%|r")
-				      end
-				      _G[battlefieldPOIName.."Timer"]:Show()
-				    else
-				      _G[battlefieldPOIName.."TextureBGTop1"]:Hide()
-				      _G[battlefieldPOIName.."TextureBGTop2"]:Hide()
-				      _G[battlefieldPOIName.."Timer"]:Hide()
-				    end
+					_G[battlefieldPOIName.."TextureBGTop1"]:Hide()
+					_G[battlefieldPOIName.."TextureBGTop2"]:Hide()
+					_G[battlefieldPOIName.."Timer"]:Hide()
 				  else
 				    local timeLeft = TIMER:TimeLeft(v.timer)
 				    _G[battlefieldPOIName.."TextureBG"]:SetWidth(RE.POIIconSize - ((timeLeft / RE.DefaultTimer) * RE.POIIconSize))
@@ -1199,9 +987,8 @@ function RE:UnitOnEnterPlayer(self, tooltipFrame)
 	for unit in pairs(self.currentMouseOverUnits) do
 		if not self:IsMouseOverUnitExcluded(unit) then
 			local unitName = UnitName(unit)
-			local unitHealth = (UnitHealth(unit) / UnitHealthMax(unit)) * 100
 			local _, _, _, unitColor = GetClassColor(select(2, UnitClass(unit)))
-			tooltipText = tooltipText..prefix.."|c"..unitColor..unitName.."|r |cFFFFFFFF["..RE:Round(unitHealth, 0).."%]|r"
+			tooltipText = tooltipText..prefix.."|c"..unitColor..unitName.."|r"
 			prefix = "\n"
 		end
 	end
@@ -1219,7 +1006,7 @@ function RE:UnitOnEnterPOI(self)
 	local tooltipText = ""
 	local battlefieldPOI = _G[self:GetName()]
 
-	if RE.CurrentMap == CI or (RE.CurrentMap == BFW and Contains(RE.BFWWalls, RE.POINodes[battlefieldPOI.name].texture)) then
+	if RE.CurrentMap == CI or (RE.CurrentMap == BFW and tContains(RE.BFWWalls, RE.POINodes[battlefieldPOI.name].texture)) then
 		return
 	end
 
@@ -1227,13 +1014,6 @@ function RE:UnitOnEnterPOI(self)
 		local status = ""
 		if RE.POINodes[battlefieldPOI.name].status and RE.POINodes[battlefieldPOI.name].status ~= "" then
 			status = "\n"..RE.POINodes[battlefieldPOI.name].status
-		end
-		if RE.POINodes[battlefieldPOI.name].health then
-			if RE.PlayedFromStart then
-				status = "\n["..RE:Round((RE.POINodes[battlefieldPOI.name].health/RE.POINodes[battlefieldPOI.name].maxHealth)*100, 0).."%]"
-			else
-				status = "\n[|r|cFFFF141D"..RE:Round((RE.POINodes[battlefieldPOI.name].health/RE.POINodes[battlefieldPOI.name].maxHealth)*100, 0).."%|r|cFFFFFFFF]"
-			end
 		end
 		if TIMER:TimeLeft(RE.POINodes[battlefieldPOI.name].timer) == 0 then
 			tooltipText = tooltipText..battlefieldPOI.name.."|cFFFFFFFF"..status.."|r"
@@ -1251,16 +1031,6 @@ function RE:UnitOnEnterPOI(self)
 		GameTooltip:Hide()
 	end
 end
-
-function RE:OnClickPOI(self)
-	if RE.CurrentMap == CI or (RE.CurrentMap == BFW and Contains(RE.BFWWalls, RE.POINodes[self.name].texture)) then
-		return
-	end
-
-	CloseDropDownMenus()
-	RE.ClickedPOI = RE.POINodes[self.name].name
-	MenuUtil.CreateContextMenu(UIParent, RE.POIDropDown)
-end
 ---
 
 -- *** Core functions
@@ -1273,7 +1043,6 @@ function RE:Startup()
 	if RE.Settings.profile.HideMinimap then
 		MinimapCluster:Hide()
 	end
-	SendAddonMessage("REPorter", "Version;"..RE.AddonVersionCheck, "INSTANCE_CHAT")
 	if IsInGuild() then
 		SendAddonMessage("REPorter", "Version;"..RE.AddonVersionCheck, "GUILD")
 	end
@@ -1289,16 +1058,13 @@ function RE:Shutdown()
 	RE.IsRated = false
 	RE.CareAboutNodes = false
 	RE.CareAboutPoints = false
-	RE.CareAboutGates = false
 	RE.CareAboutFlags = false
 	RE.CareAboutVehicles = false
-	REPorterFrame:UnregisterEvent("COMBAT_LOG_EVENT_UNFILTERED")
 	REPorterFrame:UnregisterEvent("VIGNETTES_UPDATED")
 	REPorterFrame:UnregisterEvent("AREA_POIS_UPDATED")
 	REPorterFrame:UnregisterEvent("UPDATE_UI_WIDGET")
 	REPorterFrameEstimatorText:SetText("")
 	REPorterFrameCoreUP:ResetCurrentMouseOverUnits()
-	CloseDropDownMenus()
 	if not MinimapCluster:IsShown() and RE.Settings.profile.HideMinimap then
 		MinimapCluster:Show()
 	end
@@ -1327,12 +1093,7 @@ function RE:Create()
 	REPorterFrameEstimator:SetPoint("TOP", UIWidgetTopCenterContainerFrame, "BOTTOM", 0, -10)
 	RE.POINodes = {}
 
-	if RE.CurrentMap == IOC then
-		RE.IoCGateEstimator = {}
-		RE.IoCGateEstimator[FACTION_ALLIANCE] = RE.IoCGateHealth
-		RE.IoCGateEstimator[FACTION_HORDE] = RE.IoCGateHealth
-		RE.IoCGateEstimatorText = ""
-	elseif RE.CurrentMap == SM or RE.CurrentMap == DR then
+	if RE.CurrentMap == SM or RE.CurrentMap == DR then
 		RE.SMEstimatorText = ""
 		RE.SMEstimatorReport = ""
 	else
@@ -1340,7 +1101,7 @@ function RE:Create()
 		RE.EstimatorData = {0, 0, 0, 0, -1}
 	end
 
-	if Contains({AV, BFG, IOC, AB, DG, SS, EOTS, BFW, CI, ASH, TOK}, RE.CurrentMap) then
+	if tContains({AV, BFG, IOC, AB, DG, SS, EOTS, BFW, CI, ASH, TOK}, RE.CurrentMap) then
 		RE.CareAboutNodes = true
 		if RE.CurrentMap == SS then
 			REPorterFrame:RegisterEvent("VIGNETTES_UPDATED")
@@ -1350,31 +1111,24 @@ function RE:Create()
 	else
 		RE.CareAboutNodes = false
 	end
-	if Contains({BFG, EOTS, AB, DG, SM, DR}, RE.CurrentMap) then
+	if tContains({BFG, EOTS, AB, DG, SM, DR}, RE.CurrentMap) then
 		RE.CareAboutPoints = true
 		REPorterFrame:RegisterEvent("UPDATE_UI_WIDGET")
 	else
 		RE.CareAboutPoints = false
 	end
-	if RE.CurrentMap == IOC then
-		RE.CareAboutGates = true
-		REPorterFrame:RegisterEvent("COMBAT_LOG_EVENT_UNFILTERED")
-	else
-		RE.CareAboutGates = false
-	end
-	if Contains({WG, TP, EOTS, TOK, CI, DR}, RE.CurrentMap) then
+	if tContains({WG, TP, EOTS, TOK, CI, DR}, RE.CurrentMap) then
 		RE.CareAboutFlags = true
 	else
 		RE.CareAboutFlags = false
 	end
-	if Contains({IOC, SM, BFW, DR}, RE.CurrentMap) then
+	if tContains({IOC, SM, BFW, DR}, RE.CurrentMap) then
 		RE.CareAboutVehicles = true
 	else
 		RE.CareAboutVehicles = false
 	end
 
 	RE:LoadMapSettings()
-	RE:SetupReportBar()
 	TIMER:ScheduleTimer(RE.TimerJoinCheck, 5)
 	REPorterFrameCore:SetScript("OnUpdate", RE.OnUpdate)
 end
@@ -1386,136 +1140,12 @@ function RE:NodeChange(newTexture, nodeName)
 		RE.POINodes[nodeName].isCapturing = RE.POICaptureStatus[newTexture]
 	end
 end
-
-function RE:POIStatus(POIName)
-	if RE.POINodes[POIName]then
-		if TIMER:TimeLeft(RE.POINodes[POIName].timer) == 0 then
-			if RE.POINodes[POIName].health and RE.PlayedFromStart then
-				local gateHealth = RE:Round((RE.POINodes[POIName].health / RE.POINodes[POIName].maxHealth) * 100, 0)
-				return " - Health: "..gateHealth.."%"
-			end
-			return ""
-		else
-			local timeLeft = RE:ShortTime(RE:Round(TIMER:TimeLeft(RE.POINodes[POIName].timer), 0))
-			return " - "..timeLeft
-		end
-	end
-	return ""
-end
-
-function RE:POIOwner(POIName, isReport)
-	local prefix = " - "
-	if isReport then
-		prefix = ""
-	end
-	local TranslatedName = LBS[POIName] or POIName
-	if RE.POINodes[POIName] then
-		if RE.POINodes[POIName].translatedName then
-			TranslatedName = RE.POINodes[POIName].translatedName
-		end
-		if strfind(RE.POINodes[POIName].status, FACTION_HORDE) then
-			return prefix..TranslatedName.." (Horde)"
-		elseif strfind(RE.POINodes[POIName].status, FACTION_ALLIANCE) then
-			return prefix..TranslatedName.." (Alliance)"
-		else
-			if RE.POINodes[POIName].isCapturing == FACTION_HORDE and TIMER:TimeLeft(RE.POINodes[POIName].timer) ~= 0 then
-				return prefix..TranslatedName.." (Horde)"
-			elseif RE.POINodes[POIName].isCapturing == FACTION_ALLIANCE and TIMER:TimeLeft(RE.POINodes[POIName].timer) ~= 0 then
-				return prefix..TranslatedName.." (Alliance)"
-			else
-				return prefix..TranslatedName
-			end
-		end
-	end
-	return prefix..TranslatedName
-end
-
-function RE:SmallButton(number, otherNode)
-	if select(2, IsInInstance()) == "pvp" then
-		local name
-		if otherNode then
-			name = RE.ClickedPOI
-		elseif RE.ZonesWithoutSubZones[RE.CurrentMap] then
-			name = ""
-		else
-			name = GetSubZoneText()
-		end
-		local message
-		if name and name ~= "" then
-			if number < 6 then
-				message = strupper("Incoming").." "..number
-			else
-				message = strupper("Incoming").." 5+"
-			end
-			message = message..RE:POIOwner(name)..RE:POIStatus(name)
-			SendChatMessage(message, "INSTANCE_CHAT")
-			if RE.Settings.profile.UseRaidWarnings then
-				SendChatMessage(message, "RAID_WARNING")
-			end
-		else
-			print("\124cFF74D06C[REPorter]\124r "..L["This location does not have a name. Action canceled."])
-		end
-	else
-		print("\124cFF74D06C[REPorter]\124r "..L["This addon only works in battlegrounds."])
-	end
-end
-
-function RE:BigButton(isHelp, otherNode)
-	if select(2, IsInInstance()) == "pvp" then
-		local name
-		if otherNode then
-			name = RE.ClickedPOI
-		elseif RE.ZonesWithoutSubZones[RE.CurrentMap] then
-			name = ""
-		else
-			name = GetSubZoneText()
-		end
-		local message
-		if name and name ~= "" then
-			if isHelp then
-				message = strupper("Help")..RE:POIOwner(name)..RE:POIStatus(name)
-			else
-				message = strupper("Clear")..RE:POIOwner(name)..RE:POIStatus(name)
-			end
-			SendChatMessage(message, "INSTANCE_CHAT")
-			if RE.Settings.profile.UseRaidWarnings then
-				SendChatMessage(message, "RAID_WARNING")
-			end
-		else
-			print("\124cFF74D06C[REPorter]\124r "..L["This location does not have a name. Action canceled."])
-		end
-	else
-		print("\124cFF74D06C[REPorter]\124r "..L["This addon only works in battlegrounds."])
-	end
-end
-
-function RE:ReportEstimator()
-	if IsShiftKeyDown() then
-		if TIMER:TimeLeft(RE.EstimatorTimer) > 0 then
-			SendChatMessage(RE.IsWinning.." victory: "..RE:ShortTime(RE:Round(TIMER:TimeLeft(RE.EstimatorTimer), 0)), "INSTANCE_CHAT")
-		elseif RE.CurrentMap == SM and RE.SMEstimatorReport ~= "" then
-			SendChatMessage(RE.SMEstimatorReport, "INSTANCE_CHAT")
-		elseif RE.CurrentMap == IOC and RE.PlayedFromStart then
-			SendChatMessage("Alliance gate: "..RE:Round((RE.IoCGateEstimator[FACTION_ALLIANCE] / RE.IoCGateHealth) * 100, 0).."% - Horde gate: "..RE:Round((RE.IoCGateEstimator[FACTION_HORDE] / RE.IoCGateHealth) * 100, 0).."%", "INSTANCE_CHAT")
-		end
-	end
-end
-
-function RE:ReportDropDownClick(reportType)
-	if reportType ~= "" then
-		SendChatMessage(strupper(reportType)..RE:POIOwner(RE.ClickedPOI)..RE:POIStatus(RE.ClickedPOI), "INSTANCE_CHAT")
-	else
-		SendChatMessage(RE:POIOwner(RE.ClickedPOI, true)..RE:POIStatus(RE.ClickedPOI), "INSTANCE_CHAT")
-	end
-end
 --
 
 -- *** Config functions
 function RE:UpdateConfig()
 	REPorterFrame:SetAlpha(RE.Settings.profile.Opacity)
-	REPorterBar:SetAlpha(0.25)
 	REPorterFrameBorderResize:SetShown(not RE.Settings.profile.Locked)
-	RE:SetupReportBar()
 	if select(2, IsInInstance()) == "pvp" then
 		MinimapCluster:SetShown(not RE.Settings.profile.HideMinimap)
 	end
@@ -1533,74 +1163,6 @@ function RE:UpdateScaleConfig(_, val)
 		return RE:Round(REPorterFrameCore:GetScale(), 2)
 	else
 		return 1.0
-	end
-end
-
-function RE:SetupReportBar()
-	local previousButton = "B1"
-	local handle = RE.Settings.profile.BarHandle
-	local offset = 0
-
-	if RE.IsSkinned then
-		if handle == 3 or handle == 6 or handle == 9 or handle == 12 then
-			offset = -2
-		elseif handle == 1 or handle == 4 or handle == 7 or handle == 10 then
-			offset = 2
-		end
-	end
-
-	REPorterBar:ClearAllPoints()
-	if handle < 15 and not RE.ZonesWithoutSubZones[RE.CurrentMap] then
-		REPorterBar:SetAlpha(0.25)
-		if RE.IsSkinned then
-			REPorterBarB1:SetPoint("TOPLEFT", "REPorterBar", "TOPLEFT", 5, -5)
-		else
-			REPorterBarB1:SetPoint("TOPLEFT", "REPorterBar", "TOPLEFT", 10, -10)
-		end
-		if handle < 7 or handle == 13 then
-			if handle == 13 then
-				REPorterBar:SetPoint("CENTER", "UIParent", "BOTTOMLEFT", RE.Settings.profile.BarX, RE.Settings.profile.BarY)
-			elseif handle > 3 then
-				REPorterBar:SetPoint(RE.ReportBarAnchor[handle][1], REPorterFrameBorder, RE.ReportBarAnchor[handle][2], 1, offset)
-			else
-				REPorterBar:SetPoint(RE.ReportBarAnchor[handle][1], REPorterFrameBorder, RE.ReportBarAnchor[handle][2], -1, offset)
-			end
-			if RE.IsSkinned then
-				REPorterBar:SetHeight(210)
-				REPorterBar:SetWidth(35)
-			else
-				REPorterBar:SetHeight(220)
-				REPorterBar:SetWidth(45)
-			end
-			for _, i in pairs({"B2", "B3", "B4", "B5", "B6", "B7", "B8"}) do
-				_G["REPorterBar"..i]:ClearAllPoints()
-				_G["REPorterBar"..i]:SetPoint("TOP", "REPorterBar"..previousButton, "BOTTOM")
-				previousButton = i
-			end
-		else
-			if handle == 14 then
-				REPorterBar:SetPoint("CENTER", "UIParent", "BOTTOMLEFT", RE.Settings.profile.BarX, RE.Settings.profile.BarY)
-			elseif handle > 9 then
-				REPorterBar:SetPoint(RE.ReportBarAnchor[handle][1], REPorterFrameBorder, RE.ReportBarAnchor[handle][2], offset, -1)
-			else
-				REPorterBar:SetPoint(RE.ReportBarAnchor[handle][1], REPorterFrameBorder, RE.ReportBarAnchor[handle][2], offset, 1)
-			end
-			if RE.IsSkinned then
-				REPorterBar:SetHeight(35)
-				REPorterBar:SetWidth(210)
-			else
-				REPorterBar:SetHeight(45)
-				REPorterBar:SetWidth(220)
-			end
-			for _, i in pairs({"B2", "B3", "B4", "B5", "B6", "B7", "B8"}) do
-				_G["REPorterBar"..i]:ClearAllPoints()
-				_G["REPorterBar"..i]:SetPoint("LEFT", "REPorterBar"..previousButton, "RIGHT")
-				previousButton = i
-			end
-		end
-		REPorterBar:Show()
-	else
-		REPorterBar:Hide()
 	end
 end
 
@@ -1628,17 +1190,13 @@ function RE:LoadMapSettings()
 			local currentMap = GetBestMapForUnit("player")
 			if currentMap == ABJ then
 				textures = GetMapArtLayerTextures(AB, 1)
-				RE.ZonesWithoutSubZones[AB] = true
 			elseif currentMap == ABW then
 				textures = GetMapArtLayerTextures(ABW, 1)
-				RE.ZonesWithoutSubZones[AB] = true
 			else
 				textures = GetMapArtLayerTextures(RE.CurrentMap, 1)
-				RE.ZonesWithoutSubZones[AB] = nil
 			end
 		else
 			textures = GetMapArtLayerTextures(RE.CurrentMap, 1)
-			RE.ZonesWithoutSubZones[AB] = nil
 		end
 		for i=1, #textures do
 			_G["REPorterFrameCoreMap"..i]:SetTexture(textures[i])
@@ -1667,7 +1225,6 @@ function RE:ShowDummyMap(mapID)
 
 	RE.CurrentMap = mapID
 	RE:LoadMapSettings()
-	RE:SetupReportBar()
 	REPorterFrame:Show()
 	REPorterFrame:SetFrameStrata("HIGH")
 end
@@ -1679,7 +1236,6 @@ function RE:HideDummyMap(save)
 		RE.LastMap = 0
 		REPorterFrame:Hide()
 		REPorterFrame:SetFrameStrata("MEDIUM")
-		REPorterBar:Hide()
 	end
 end
 --