+++ AddOns/REPorter/CHANGELOG.md
@@ -1,6 +1,11 @@
 # REPorter
 
-## [v2.10.11](https://github.com/AcidWeb/REPorter/tree/v2.10.11) (2025-08-21)
-[Full Changelog](https://github.com/AcidWeb/REPorter/compare/v2.10.10...v2.10.11) [Previous Releases](https://github.com/AcidWeb/REPorter/releases)
+## [v3.0.0](https://github.com/AcidWeb/REPorter/tree/v3.0.0) (2026-02-07)
+[Full Changelog](https://github.com/AcidWeb/REPorter/compare/v2.10.11...v3.0.0) [Previous Releases](https://github.com/AcidWeb/REPorter/releases)
 
-- TOC bump  
+- Miscellaneous tweaks  
+- Code cleanup  
+- Added new ElvUI skin  
+- Removed player health tracking  
+- Removed IoC gate health tracking  
+- Removed reporting